Move MarkdownRenderer into `platform` layer
For #206228

This moves the MarkdownRendererService from `editor` into `platform`

As part of this I had to add a `setDefaultCodeBlockRenderer` method that lets the editor register syntax highlighting support
