+++ src/vs/workbench/contrib/comments/browser/commentThreadZoneWidget.ts
@@ -271,7 +271,7 @@ export class ReviewZoneWidget extends ZoneWidget implements ICommentThreadWidget
 			this._commentThread,
 			this._pendingComment,
 			this._pendingEdits,
-			{ editor: this.editor, },
+			{ context: this.editor, },
 			this._commentOptions,
 			{
 				actionRunner: async () => {
